Renamed prepNextDevRelease to prepBetaRelease (#1193)
* Renamed prepNextDevRelease to prepBetaRelease

* Removed unnecessary console.log

* Renaming next dev functions to beta functions

* Fixed comments to reflect the new changes

* Fixed dev check

* Change files

* Changed changelog message to past tense

* Modified yarn beachball execution to execute from the root of the monorepo instead

* Removed unused functions,added new getNextPrefix function, removed references to next-dev

* Added space back to // comments for consistency, renamed getNextPrefix to getNewerPrefix, and removed unnecessary code

* Fixed reference to getNewerPrefix
