+++ change/@microsoft-teams-js-b469c20c-b2ae-4fcb-aa82-4d19b3292ae9.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "Replaced prepNextDevRelease.js with prepBetaRelease.js to handle beta version bumps and deprecated use of nextDev tags for releases",
+  "packageName": "@microsoft/teams-js",
+  "email": "jdahiya8719@outlook.com",
+  "dependentChangeType": "none"
+}
